content-augmentation/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud & Firebase
google-cloud-storage==2.14.0
google-cloud-functions==1.4.0
google-cloud-firestore==2.14.0
google-auth==2.27.0
google-api-python-client==2.116.0

# Gemini AI
google-generativeai==0.3.2

# Task Queue & Workers (Optional)
celery==5.3.6
redis==5.0.1

# HTTP & API Clients
requests==2.31.0
httpx==0.26.0
aiohttp>=3.9.0

# PDF & Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Image Processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
structlog==24.1.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
